Add new folder for pages: include Dashboard, Signupf1, and Signupf2
